Install and bower only if bower.json is present

+++ lib/language_pack/ruby.rb
@@ -580,17 +580,20 @@ ERROR
 
   # install bower as npm module
   def install_bower
-    log("bower") do
-      run("curl #{BOWER_BASE_URL}/bower-#{BOWER_VERSION}/node_modules.tar.gz -s -o - | tar xzf -")
-      unless $?.success?
-        error "Can't install Bower #{BOWER_VERSION}"
+    if File.exist?('bower.json')
+      log("bower") do
+        run("curl #{BOWER_BASE_URL}/bower-#{BOWER_VERSION}/node_modules.tar.gz -s -o - | tar xzf -")
+        unless $?.success?
+          error "Can't install Bower #{BOWER_VERSION}"
+        end
       end
     end
   end
 
   # runs bower to install the dependencies
   def build_bower
-    error_message = <<ERROR
+    if File.exist?('bower.json')
+      error_message = <<ERROR
 Can't install JavaScript dependencies
 
 Bower 1.0.0 released at 2013-07-23
@@ -601,19 +604,20 @@ Check these points:
 * bower.json requires 'name' option
 ERROR
 
-    log("bower") do
-      topic("Installing JavaScript dependencies using Bower #{BOWER_VERSION}")
+      log("bower") do
+        topic("Installing JavaScript dependencies using Bower #{BOWER_VERSION}")
 
-      load_bower_cache
+        load_bower_cache
 
-      pipe("./node_modules/bower/bin/bower install --config.storage.packages=vendor/bower/packages --config.storage.registry=vendor/bower/registry --config.tmp=vendor/bower/tmp 2>&1")
-      if $?.success?
-        log "bower", :status => "success"
-        puts "Cleaning up the Bower tmp."
-        FileUtils.rm_rf("vendor/bower/tmp")
-        cache.store "vendor/bower"
-      else
-        error error_message
+        pipe("./node_modules/bower/bin/bower install --config.storage.packages=vendor/bower/packages --config.storage.registry=vendor/bower/registry --config.tmp=vendor/bower/tmp 2>&1")
+        if $?.success?
+          log "bower", :status => "success"
+          puts "Cleaning up the Bower tmp."
+          FileUtils.rm_rf("vendor/bower/tmp")
+          cache.store "vendor/bower"
+        else
+          error error_message
+        end
       end
     end
   end
